move flow cell encryption api
